@@ -1,4 +1,4 @@
-import { BrowserRouter, Route, Routes } from "react-router-dom";
+import { Route, Routes } from "react-router-dom";
 
 import Landing from "./components/Landing";
 import Listings from "./components/Listings";
@@ -7,12 +7,10 @@ import "./App.css";
 
 function App() {
   return (
-    <BrowserRouter>
-      <Routes>
-        <Route path="/listings" element={<Listings />} />
-        <Route path="/" element={<Landing />} />
-      </Routes>
-    </BrowserRouter>
+    <Routes>
+      <Route path="/listings" element={<Listings />} />
+      <Route path="/" element={<Landing />} />
+    </Routes>
   );
 }
 
